pyproject.toml
[tool.poetry]
name = "djapiplayground"
version = "0.1.0"
description = "Django API Playground"
authors = ["Louis Agyapong <47106199+louis-agyapong@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
python-decouple = "^3.6"
django-extensions = "^3.2.1"
ipython = "^8.8.0"
dj-database-url = "^1.2.0"
djangorestframework = "^3.14.0"
black = "^22.12.0"
mypy = "^0.991"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
